+++ Engine/Main/Application.cpp
@@ -15,7 +15,7 @@
 #include "Module/ModuleUI.h"
 #include "Module/ModuleFileSystem.h"
 #include "Module/ModuleWindow.h"
-#include "ResourceManagement/Importer/MaterialImporter.h"
+#include "ResourceManagement/Importer/TextureImporter.h"
 #include "ResourceManagement/Importer/ModelImporter.h"
 #include "UI/EngineLog.h"
 #include "Brofiler/Brofiler.h"
