+++ Engine/Module/ModuleModelLoader.cpp
@@ -11,7 +11,7 @@
 #include "Component/ComponentMaterial.h"
 #include "Component/ComponentMesh.h"
 #include "ResourceManagement/Importer/ModelImporter.h"
-#include "ResourceManagement/Importer/MaterialImporter.h"
+#include "ResourceManagement/Importer/TextureImporter.h"
 
 #include <assimp/cimport.h>
 #include <assimp/postprocess.h>
