+++ Engine/ResourceManagement/Importer/TextureImporter.cpp
@@ -1,4 +1,4 @@
-#include "MaterialImporter.h"
+#include "TextureImporter.h"
 #include "Main/Application.h"
 
 #include <algorithm>
